+++ client/web/src/organisms/mainpage/featureSuggestion/FeatureDetail.tsx
@@ -71,11 +71,7 @@ export default function FeatureDetail({
 
   // Next Feature
   const nextFeature = data[currentSuggestionIndex + 1];
-  // const handleEditButton = (id: number) => {
-  //   <Link to={`/feature-suggestion/${id}/edit`} />;
-  //   // window.location.replace(`/feature-suggestion/${id}/edit`);
-  //   // setEditState(true);
-  // };
+
   const doDelete = () => {
     const doConfirm = window.confirm('삭제 하시겠습니까?');
     if (doConfirm) {
@@ -85,132 +81,118 @@ export default function FeatureDetail({
   };
   return (
     <div>
-      {!editState ? (
-        <div>
-          <Paper component="article">
-            <div className={classes.title}>
-              <Typography variant="h6" className={classes.titleText}>
-                {currentSuggestion?.title}
-              </Typography>
-              <Typography color="textSecondary">
-                {categoryTabSwitch(Number(currentSuggestion?.category))}
-                {currentSuggestion ? new Date(currentSuggestion.createdAt).toLocaleString() : ''}
-              </Typography>
-            </div>
-
-            <div className={classes.contentsText}>
-              <Markdown
-                className={classes.markdown}
-                source={currentSuggestion?.content}
-                escapeHtml={false}
-                // eslint-disable-next-line react/prop-types
-                renderers={{ code: ({ value }) => <Markdown source={value} /> }}
-              />
-            </div>
-            {currentSuggestion?.reply ? (
-              <div className={classes.replyCard}>
-                <Divider />
-                <Typography variant="h6">
-                  관리자 답글
-                </Typography>
-                <Card className={classes.replyTextCard}>
-                  <Typography variant="body1">
-                    {currentSuggestion.reply}
-                  </Typography>
-                </Card>
-              </div>
-            ) : null}
-          </Paper>
-
-          <div id="button-set" className={classes.buttonSet}>
-            <Button
-              style={{ width: '25%' }}
-              size="large"
-              disabled={currentSuggestionIndex === 0}
-              variant="contained"
-              color="primary"
-              onClick={() => {
-                onOtherFeatureClick(previousFeature.id);
-              }}
-            >
-              <KeyboardArrowLeft />
-              {currentSuggestionIndex !== 0 && (
-                <Typography>
-                  {previousFeature.title.length > TITLE_LENGTH
-                    ? `${previousFeature.title.slice(0, TITLE_LENGTH)}...`
-                    : previousFeature.title}
-                </Typography>
-              )}
-            </Button>
-            <Button
-              style={{ width: '10%' }}
-              size="large"
-              variant="contained"
-              onClick={() => {
-                onBackClick();
-              }}
-            >
-              목록으로
-            </Button>
-            <Button
-              style={{ width: '25%' }}
-              size="large"
-              disabled={currentSuggestionIndex === data.length - 1}
-              variant="contained"
-              color="primary"
-              onClick={() => {
-                onOtherFeatureClick(nextFeature.id);
-              }}
-            >
-              {currentSuggestionIndex !== data.length - 1 && (
-                <Typography>
-                  {nextFeature.title.length > TITLE_LENGTH
-                    ? `${nextFeature.title.slice(0, TITLE_LENGTH)}...`
-                    : nextFeature.title}
-                </Typography>
-              )}
-              <KeyboardArrowRight />
-            </Button>
+      <div>
+        <Paper component="article">
+          <div className={classes.title}>
+            <Typography variant="h6" className={classes.titleText}>
+              {currentSuggestion?.title}
+            </Typography>
+            <Typography color="textSecondary">
+              {categoryTabSwitch(Number(currentSuggestion?.category))}
+              {currentSuggestion ? new Date(currentSuggestion.createdAt).toLocaleString() : ''}
+            </Typography>
           </div>
-          {currentSuggestion?.author === authContext.user.userId
-            ? (
-              <ButtonGroup disableElevation variant="contained" color="primary">
-                <Link 
-                  to={{
-                    pathname: `/feature-suggestion/write/${currentSuggestion?.id}`,
-                    state: [currentSuggestion],
-                  }}
-                >
-                  수정
-                </Link>
 
-                {/* <Button
-                  size="large"
-                  variant="contained"
-                  color="primary"
-                  onClick={() => {
-                    handleEditButton(currentSuggestion?.id);
-                  }}
-                > */}
-                {/* 수정 */}
-                {/* </Button> */}
-                <Button
-                  size="large"
-                  variant="contained"
-                  color="secondary"
-                  onClick={doDelete}
-                >
-                  삭제
-                </Button>
-              </ButtonGroup>
-            ) : null}
+          <div className={classes.contentsText}>
+            <Markdown
+              className={classes.markdown}
+              source={currentSuggestion?.content}
+              escapeHtml={false}
+                // eslint-disable-next-line react/prop-types
+              renderers={{ code: ({ value }) => <Markdown source={value} /> }}
+            />
+          </div>
+          {currentSuggestion?.reply ? (
+            <div className={classes.replyCard}>
+              <Divider />
+              <Typography variant="h6">
+                관리자 답글
+              </Typography>
+              <Card className={classes.replyTextCard}>
+                <Typography variant="body1">
+                  {currentSuggestion.reply}
+                </Typography>
+              </Card>
+            </div>
+          ) : null}
+        </Paper>
+
+        <div id="button-set" className={classes.buttonSet}>
+          <Button
+            style={{ width: '25%' }}
+            size="large"
+            disabled={currentSuggestionIndex === 0}
+            variant="contained"
+            color="primary"
+            onClick={() => {
+              onOtherFeatureClick(previousFeature.id);
+            }}
+          >
+            <KeyboardArrowLeft />
+            {currentSuggestionIndex !== 0 && (
+            <Typography>
+              {previousFeature.title.length > TITLE_LENGTH
+                ? `${previousFeature.title.slice(0, TITLE_LENGTH)}...`
+                : previousFeature.title}
+            </Typography>
+            )}
+          </Button>
+          <Button
+            style={{ width: '10%' }}
+            size="large"
+            variant="contained"
+            onClick={() => {
+              onBackClick();
+            }}
+          >
+            목록으로
+          </Button>
+          <Button
+            style={{ width: '25%' }}
+            size="large"
+            disabled={currentSuggestionIndex === data.length - 1}
+            variant="contained"
+            color="primary"
+            onClick={() => {
+              onOtherFeatureClick(nextFeature.id);
+            }}
+          >
+            {currentSuggestionIndex !== data.length - 1 && (
+            <Typography>
+              {nextFeature.title.length > TITLE_LENGTH
+                ? `${nextFeature.title.slice(0, TITLE_LENGTH)}...`
+                : nextFeature.title}
+            </Typography>
+            )}
+            <KeyboardArrowRight />
+          </Button>
         </div>
-      )
-        : (
-          <FeatureWriteForm
-            editData={currentSuggestion}
-          />
-        )}
+        {currentSuggestion?.author === authContext.user.userId
+          ? (
+            <ButtonGroup disableElevation variant="contained" color="primary">
+              <Button
+                component={Link}
+                to={{
+                  pathname: `/feature-suggestion/write/${currentSuggestion?.id}`,
+                  state: [currentSuggestion],
+                }}
+                size="large"
+                variant="contained"
+                color="primary"
+              >
+                수정
+              </Button>
+              <Button
+                size="large"
+                variant="contained"
+                color="secondary"
+                onClick={doDelete}
+              >
+                삭제
+              </Button>
+            </ButtonGroup>
+          ) : null}
+      </div>
     </div>
   );
 }
